Move ReadInputStep impl to dedicated file
